Config struct tests passed
