+++ config.toml
@@ -1,9 +1,9 @@
 # Configuration for your trading application
 
-exchange = "bybit" # Name of the exchange you're connecting to
+exchange = "bybit" # Name of the exchange you're connecting to (e.g., bybit, binance, both)
 
 # List of symbols to track and trade
-symbols = ["ETHUSDT"] # Replace with your desired symbols
+symbols = ["NOTUSDT", "DOGSUSDT"] # Replace with your desired symbols
 
 # API keys for authentication (potentially multiple keys for redundancy)
 api_keys = [
@@ -20,22 +20,19 @@ api_keys = [
 ]
 
 # Initial balances for tracked symbols (symbol, balance)
-balances = [["BTC", 0.1], ["ETH", 1.0]]
+balances = [["NOTUSDT", 100.0], ["NOTUSDT", 100.0]]
 
 # Leverage for trading positions (e.g., 5.0 for 5x leverage)
-leverage = 2.0
+leverage = 10.0
 
 # Maximum number of orders to place on each side of the market
-orders_per_side = 3
+orders_per_side = 6
 
-# Final order distance for trailing stop orders (e.g., 0.01 for 1% price difference)
-final_order_distance = 0.005
+# Final order distance for orders (e.g., 10 for 10 * 25bps = 250bps or 2.5%)
+final_order_distance = 10
 
 # Number of order book levels (depth) to request from the exchange
-depths = [10, 20] # You can have multiple depths for different symbols
-
-# Rebalance ratio for portfolio management (e.g., 0.1 for 10% rebalance)
-rebalance_ratio = 0.05
+depths = [4, 13, 32] # You can have multiple depths for different symbols
 
 # Rate limit for API calls (in milliseconds)
 rate_limit = 1000
@@ -43,5 +40,5 @@ rate_limit = 1000
 # Profit spread for  (e.g., 1 for 0.01% profit spread)
 bps = [27, 25]
 
-# use wmid when calculating 
-use_wmid = true
+# use tick window for waiting and lookback period
+tick_window = 610
